// api/src/_generated/nestjs/resolvers/inputs/UserScalarWhereWithAggregatesInput.ts
import * as NestJS from "@nestjs/graphql";
import * as GraphQLScalars from "graphql-scalars";
import { Prisma } from "@prisma/client";
import { DecimalJSScalar } from "../../scalars";
import { DateTimeWithAggregatesFilter } from "../inputs/DateTimeWithAggregatesFilter";
import { EnumUserRoleNullableWithAggregatesFilter } from "../inputs/EnumUserRoleNullableWithAggregatesFilter";
import { StringNullableWithAggregatesFilter } from "../inputs/StringNullableWithAggregatesFilter";
import { StringWithAggregatesFilter } from "../inputs/StringWithAggregatesFilter";
import { UuidNullableWithAggregatesFilter } from "../inputs/UuidNullableWithAggregatesFilter";
import { UuidWithAggregatesFilter } from "../inputs/UuidWithAggregatesFilter";

@NestJS.InputType("UserScalarWhereWithAggregatesInput", {
  isAbstract: true
})
export class UserScalarWhereWithAggregatesInput {
  @NestJS.Field(_type => [UserScalarWhereWithAggregatesInput], {
    nullable: true
  })
  AND?: UserScalarWhereWithAggregatesInput[] | undefined;

  @NestJS.Field(_type => [UserScalarWhereWithAggregatesInput], {
    nullable: true
  })
  OR?: UserScalarWhereWithAggregatesInput[] | undefined;

  @NestJS.Field(_type => [UserScalarWhereWithAggregatesInput], {
    nullable: true
  })
  NOT?: UserScalarWhereWithAggregatesInput[] | undefined;

  @NestJS.Field(_type => UuidWithAggregatesFilter, {
    nullable: true
  })
  id?: UuidWithAggregatesFilter | undefined;

  @NestJS.Field(_type => DateTimeWithAggregatesFilter, {
    nullable: true
  })
  createdAt?: DateTimeWithAggregatesFilter | undefined;

  @NestJS.Field(_type => DateTimeWithAggregatesFilter, {
    nullable: true
  })
  updatedAt?: DateTimeWithAggregatesFilter | undefined;

  @NestJS.Field(_type => StringWithAggregatesFilter, {
    nullable: true
  })
  email?: StringWithAggregatesFilter | undefined;

  @NestJS.Field(_type => StringNullableWithAggregatesFilter, {
    nullable: true
  })
  name?: StringNullableWithAggregatesFilter | undefined;

  @NestJS.Field(_type => StringWithAggregatesFilter, {
    nullable: true
  })
  imageSmall?: StringWithAggregatesFilter | undefined;

  @NestJS.Field(_type => StringWithAggregatesFilter, {
    nullable: true
  })
  imageLarge?: StringWithAggregatesFilter | undefined;

  @NestJS.Field(_type => EnumUserRoleNullableWithAggregatesFilter, {
    nullable: true
  })
  role?: EnumUserRoleNullableWithAggregatesFilter | undefined;

  @NestJS.Field(_type => UuidNullableWithAggregatesFilter, {
    nullable: true
  })
  organisationId?: UuidNullableWithAggregatesFilter | undefined;
}
